pyproject.toml
[tool.poetry]
name = "py-template"
version = "0.1.0"
description = ""
authors = ["Norbert <norbert.lipinski2@sky.uk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
polars = "^0.19.3"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
pendulum = "^2.1.2"
loguru = "^0.7.2"
lxml = "^4.9.3"
google-api-python-client = "^2.107.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.3"
pyright = "^1.1.333"
pytest = "~7.4.0"
pytest-randomly = "^3.15.0"


[tool.ruff]
target-version = "py311"

[tool.ruff.lint]
extend-select = ["RUF", "I"]

[tool.pyright]
include = ["app"]

[tool.pytest.ini_options]
testpaths = ['test']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
